fortsetter forsøk på oppryddingen. er litt å ta i. splitter bakgrunnskart i mindre objekt så det er lettere å følge logikken
