+++ src/Tweaks/Undermenyer/bakgrunn/Tema.js
@@ -1,13 +1,13 @@
 import { List } from "@material-ui/core";
 import React from "react";
-import TemaPreview from "./TemaPreview";
+import TemaButton from "./TemaButton";
 import typer from "./temaer";
 
 const Tema = ({ valgt, onUpdateLayerProp }) => (
   <List>
     {Object.keys(typer).map(key => {
       return (
-        <TemaPreview
+        <TemaButton
           button
           type={key}
           valgt={valgt}
